requirements.txt
# Data Collection
scholarly>=1.7.11
arxiv>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
semanticscholar>=0.8.0  # For citation data enrichment

# Data Processing
pandas>=2.1.0
python-dateutil>=2.8.2

# AI/LLM (Phase 2)
openai>=1.0.0
httpx>=0.25.0

# Database
sqlalchemy>=2.0.0

# Scheduling
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging
loguru>=0.7.0
